Add the directory recipes and imagesOfRecipes and html files index & shahiPaneer and image of Shahi-Paneer
By adding the recipes directory various recipes html file can be put in it and by adding imagesOfRecipes we can put recipes images in it and show it in our recipes file like shahiPaneer.html & index.html is our homepage of our website.

Adding directory so it is easy to organize our files.